xss HTML entity fix
With the sanitize option on it is possible to create a link with a javascript: protocol with the following: 
   [URL](javascript&#58document;alert&#40;1&#41;).

HTML entities in the browser are not strict and parse what they can and leaving the rest behind. For example &#xNNanything; would parse the NN hex values but leave behind the string "anything;".

"javascript&#58document;" with the regex /&([#\w]+);/ returns "58document" and is parsed by String.fromCharCode to "". Because of this the later tests only sees the javascript keyword without the :. However the browser parses this to: "javascript:document;".

from: https://github.com/chjj/marked/pull/592
Merge branch 'xss_html_entities' of github.com:matt-/marked